+++ src/NEMO_mqtt_bridge/migrations/0007_replace_tls_with_hmac.py
@@ -6,7 +6,7 @@ from django.db import migrations, models
 class Migration(migrations.Migration):
 
     dependencies = [
-        ("nemo_mqtt", "0006_rename_db_tables_to_lowercase"),
+        ("NEMO_mqtt_bridge", "0006_rename_db_tables_to_lowercase"),
     ]
 
     operations = [