+++ src/NEMO_mqtt_bridge/migrations/0009_hmac_sha256_only.py
@@ -4,14 +4,14 @@ from django.db import migrations
 
 
 def set_hmac_algorithm_sha256(apps, schema_editor):
-    MQTTConfiguration = apps.get_model("nemo_mqtt", "MQTTConfiguration")
+    MQTTConfiguration = apps.get_model("NEMO_mqtt_bridge", "MQTTConfiguration")
     MQTTConfiguration.objects.all().update(hmac_algorithm="sha256")
 
 
 class Migration(migrations.Migration):
 
     dependencies = [
-        ("nemo_mqtt", "0008_fix_qos_level"),
+        ("NEMO_mqtt_bridge", "0008_fix_qos_level"),
     ]
 
     operations = [